title was added to sites screen
