+++ features/main/src/main/res/values/strings.xml
@@ -4,6 +4,7 @@
     <string name="site">Сайт:</string>
     <string name="link">Ссылка:</string>
     <string name="password">Пароль:</string>
+    <string name="saved_passwords">Сохранённые пароли</string>
     <string name="hidden_text">*******</string>
     <string name="no_sites">У вас ещё нет сохранённых паролей</string>
     <string name="title_edit">Редактировать</string>
